+++ Stopwatch/markers_backup.txt
@@ -1,0 +1,13 @@
+1 00:00:00:00 1 0 False 
+2 00:00:00:00 1 0 False 
+3 00:00:00:00 1 0 False 
+4 00:00:00:00 1 0 False 
+5 00:00:01:31 1 0 False 
+6 00:00:02:99 1 0 False 
+7 00:00:04:09 1 0 False 
+8 00:00:05:21 1 0 False 
+9 00:00:06:11 1 0 False 
+10 00:00:07:17 1 0 False 
+11 00:00:08:69 1 0 False 
+12 00:00:09:70 1 0 False 
+13 00:00:00:00 1 0 False
